Small fixes.

+++ src/main/java/com/example/VideoMicroservice_CopyAllToNew/repositories/ArtistRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ArtistRepository extends JpaRepository<Artist, Integer> {
+public interface ArtistRepository extends JpaRepository<Artist, Long> {
     Optional<Artist> findByNameIgnoreCase(String name);
     boolean existsByNameIgnoreCase(String name);
 }
